Fix bugs and drop support for Postgres >= 9.0
`complex--0.5.0.sql` was ending up in the `DATA` variable, which causes
newer versions of `make` to error out. So simplify the `Makefile` a bit
by requiring Postgres 9.1 and removing the conditional code. Also remove
references to installing the extension on earlier versions of Postgres.

Requires Postgres 9.1 in `META.json` and fix the GitHub URLs.

Add a second expected output for updated formatting on newer versions of
Postgres (I tested with 16.2).